refactor schedule components
