+++ src/containers/Schedule/EventActivity.tsx
@@ -8,12 +8,12 @@ import {
 import { useCallback, useEffect, useMemo } from 'react';
 import { Link } from 'react-router-dom';
 import { CutoffTimeLimitPanel } from '@/components/CutoffTimeLimitPanel';
+import { PeopleList } from '@/components/PeopleList';
 import { getRooms } from '@/lib/activities';
 import { isRankedBySingle } from '@/lib/events';
 import { renderResultByEventId } from '@/lib/results';
 import { formatDateTimeRange } from '@/lib/time';
 import { useWCIF } from '@/providers/WCIFProvider';
-import { PeopleList } from './PeopleList';
 
 const isAssignment = (assignment) => (a) =>
   a.assignments.some(({ assignmentCode }) => assignmentCode === assignment);
