+++ src/containers/Schedule/index.ts
@@ -1,1 +1,6 @@
 export * from './Schedule';
+export * from './Activity';
+export * from './Room';
+export * from './Rooms';
+export * from './EventActivity';
+export * from './OtherActivity';
